Fixed > vs >= error in node collision check.
This fixes the edge collision check failure, as the current algorithm
for edge checking does not detect if the line is entirely inside
(doesn't cross the borders). Obviously eventually the edge collision
checcheck function should also check for this case. But as it is, the
isEisEdgeColliding will not work as expected, when both start and end
nodes are on the obstacle border